Add a PubTator based MCP server
Can be used as an alternative implementation to finding papers and extracting
variants associated with the gene under consideration.

The configs have been updated to use the previous LLM-based implementations,
but it's now possible to swap that with the MCP server based implementations
instead, as described in the commented-out sections of lib/config/evagg_pipeline.yaml.